Move mocks into top level of background
