+++ projects/extension/src/background/AppMediator.test.ts
@@ -1,6 +1,6 @@
 import { ExtensionMessage } from '../types';
 import { AppMediator } from './AppMediator';
-import { MockPort, MockConnectionManager } from './mocks';
+import { MockPort, MockConnectionManager } from '../mocks';
 import { JsonRpcResponse } from './types';
 
 test('initialises correctly', () => {
